refactor: move conversation template to config/
