--- apps/bixarena/app/bixarena_app/README.md
@@ -1,84 +1,0 @@
-# bixarena_app - Refactored Structure
-
-This directory contains the bixarena_app after successful decoupling from the full FastChat library and reorganization into a cleaner structure.
-
-## Current Structure
-
-```
-bixarena_app/
-├── config/                 # Configuration and utilities
-│   ├── constants.py       # Constants (merged from FastChat)
-│   ├── utils.py          # Utilities (build_logger, etc.)
-│   └── prompt_examples.py
-├── model/                  # Model-related components
-│   ├── conversation.py    # Conversation prompt templates
-│   ├── model_adapter.py   # Simplified model adapter
-│   ├── api_provider.py    # API provider streaming
-│   ├── model_response.py  # Main model response logic
-│   └── model_selection.py
-├── auth/                   # Authentication
-├── page/                   # UI pages
-└── main.py
-```
-
-## Refactoring History
-
-### Phase 1: Decoupling from FastChat
-
-- Removed the entire `fastchat/` directory (hundreds of files)
-- Created `temp_fastchat/` with minimal required components
-- Updated imports in `model_response.py`
-
-### Phase 2: Reorganization (Current)
-
-- **Merged constants**: Combined `temp_fastchat/constants.py` with existing `config/constants.py`
-- **Added utils**: Created `config/utils.py` with `build_logger` function
-- **Moved core files to model/**:
-  - `conversation.py` → `model/conversation.py`
-  - `api_provider.py` → `model/api_provider.py`
-  - `model_adapter.py` → `model/model_adapter.py`
-- **Updated all imports** to use new paths
-- **Removed temp_fastchat** directory completely
-
-## What was simplified
-
-The model adapter was greatly simplified to only include:
-
-- Basic model adapter interface
-- Support for OpenAI (GPT), Anthropic (Claude), and Google (Gemini) models
-- Conversation template functionality
-- Removed heavy dependencies on torch, transformers, and other ML libraries
-
-## Current Dependencies
-
-After refactoring, bixarena_app only depends on:
-
-- Core conversation templates and API providers
-- Simplified logging utilities
-- Basic model adapters for API-based models
-
-## Key Benefits
-
-1. **Massive size reduction**: Removed hundreds of unused FastChat files
-2. **Clean organization**: Related functionality grouped logically
-3. **Simplified imports**: Clear import paths within bixarena_app
-4. **Maintained functionality**: All existing features preserved
-5. **Better maintainability**: Easier to understand and modify
-
-## Import Usage
-
-The main file that uses these components is `model/model_response.py`:
-
-```python
-from bixarena_app.config.constants import ErrorCode, SERVER_ERROR_MSG
-from bixarena_app.model.model_adapter import get_conversation_template
-from bixarena_app.model.api_provider import get_api_provider_stream_iter
-```
-
-## Future Improvements
-
-The current structure is much cleaner and ready for further refinements:
-
-1. Consider consolidating conversation templates
-2. Optimize API provider logic if needed
-3. Continue removing unused functionality
